Added day 2 solution

+++ 2025/Day_01.cs
@@ -2,7 +2,7 @@ namespace aoc;
 
 public static class Day_01
 {
-    public static int Part1(SolutionTimer timer, string[] input)
+    public static long Part1(SolutionTimer timer, string[] input)
     {
         timer.StartParsing();
 
@@ -35,7 +35,7 @@ public static class Day_01
     }
     
 
-    public static int Part2(SolutionTimer timer, string[] input)
+    public static long Part2(SolutionTimer timer, string[] input)
     {
         timer.StartParsing();
